made some changes to urls and article models
