+++ app/article/views.py
@@ -1,3 +1,8 @@
-from django.shortcuts import render
+from rest_framework import viewsets
+from .models import Article
+from .serializers import ArticleSerializer
 
-# Create your views here.
+
+class ArticleList(viewsets.ModelViewSet):
+    queryset = Article.objects.all()
+    serializer_class = ArticleSerializer
