+++ app/core/urls.py
@@ -26,6 +26,6 @@ urlpatterns = [
     path('ht/', include('health_check.urls')),
     path('__debug__/', include('debug_toolbar.urls')),
     path('api/v1/auth/', include('user.urls')),
-    path('medium/', include('content.api.urls')),
-    path('article/', include('article.api.urls')),
+    path('content/', include('content.urls')),
+    path('article/', include('article.urls')),
 ]
